+++ packages/elements/src/photon-multirx-form/prescribe-workflow-analytics.test.tsx
@@ -11,61 +11,11 @@ import {
 } from './photon-prescribe-workflow-component';
 import { createTestClient, createTestClientStore } from '../test-utils/createTestClient';
 import { clinicalGql, defaultHandlers, DISPENSE_UNIT, TREATMENT } from '../test-utils/msw-handlers';
+import { MockMedicationSearchElement } from '../test-utils/mock-medication-search.element';
 
 vi.mock('solid-element', () => ({
   customElement: vi.fn()
 }));
-
-// todo: remove this MockMedicationSearchElement
-class MockMedicationSearchElement extends HTMLElement {
-  connectedCallback() {
-    if (this.dataset.initialized === 'true') return;
-    this.dataset.initialized = 'true';
-
-    const label = this.getAttribute('label') || 'Search for Treatment';
-    const wrapper = document.createElement('label');
-    wrapper.textContent = label;
-
-    const select = document.createElement('select');
-    select.setAttribute('aria-label', label);
-
-    const placeholder = document.createElement('option');
-    placeholder.value = '';
-    placeholder.textContent = 'Select treatment';
-    select.append(placeholder);
-
-    const option = document.createElement('option');
-    option.value = TREATMENT.id;
-    option.textContent = TREATMENT.name;
-    select.append(option);
-
-    select.addEventListener('change', () => {
-      if (!select.value) {
-        this.dispatchEvent(
-          new CustomEvent('photon-treatment-unselected', { bubbles: true, composed: true })
-        );
-        return;
-      }
-      this.dispatchEvent(
-        new CustomEvent('photon-search-text-changed', {
-          bubbles: true,
-          composed: true,
-          detail: { text: TREATMENT.name }
-        })
-      );
-      this.dispatchEvent(
-        new CustomEvent('photon-treatment-selected', {
-          bubbles: true,
-          composed: true,
-          detail: { data: TREATMENT, catalogId: 'cat_123' }
-        })
-      );
-    });
-
-    wrapper.append(select);
-    this.append(wrapper);
-  }
-}
 
 const server = setupServer(...defaultHandlers);
 
@@ -252,7 +202,7 @@ test('CTAs', async () => {
   );
 });
 
-test('attestation CTAs — Attestation Canceled', async () => {
+test('attestation CTA', async () => {
   const { analyticsEvents, user } = renderPrescribeWorkflow({}, { attestationStatus: 'NEEDS' });
 
   await waitForSignatureAttestationModal();
@@ -286,8 +236,8 @@ function renderPrescribeWorkflow(
   props: Partial<PrescribeWorkflowComponentProps> = {},
   options: { attestationStatus?: 'COMPLETE' | 'NEEDS' } = {}
 ) {
-  // Per-test MSW overrides
   if (options.attestationStatus === 'NEEDS') {
+    // Per-test MSW overrides
     server.use(
       clinicalGql.query('GetCurrentUserSignatureAttestationStatus', () =>
         HttpResponse.json({
